Created Controller class for event creation

+++ src/main/java/com/example/webgrow/repository/EventRepository.java
@@ -3,5 +3,5 @@ import com.example.webgrow.models.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 public interface EventRepository extends JpaRepository<Event, Long> {
-    List<Event> findByHostId(Long hostId);
+    List<Event> findByHostId(Integer hostId);
 }
